project/requirements.txt
absl-py==0.12.0
astunparse==1.6.3
asyncqt==0.8.0
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
decorator==5.0.9
dill==0.3.3
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
importlib-resources==5.1.4
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
matplotlib-inline==0.1.2
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.2.52
opt-einsum==3.3.0
pandas==1.2.4
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
Pygments==2.9.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
sklearn==0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-addons==0.13.0
tensorflow-datasets==4.3.0
tensorflow-estimator==2.5.0
tensorflow-metadata==1.0.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
typeguard==2.12.0
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
